move auth DTOs to separate directory
