+++ src/main/java/org/mae/twg/backend/controllers/auth/AuthAdminController.java
@@ -4,8 +4,8 @@ import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
-import org.mae.twg.backend.dto.SignInRequest;
-import org.mae.twg.backend.dto.SignUpRequest;
+import org.mae.twg.backend.dto.auth.SignInRequest;
+import org.mae.twg.backend.dto.auth.SignUpRequest;
 import org.mae.twg.backend.services.auth.AuthAdminService;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PostMapping;
